Fix warning C4244 on MSVC.

+++ src/core/text.cpp
@@ -12,13 +12,17 @@ void set_integer_position(sf::Text &text,
                           const float x,
                           const float y)
 {
-    text.setPosition(static_cast<int>(x), static_cast<int>(y));
+    // Cast to integer, then back to float
+    text.setPosition(static_cast<float>(static_cast<int>(x)),
+                     static_cast<float>(static_cast<int>(y)));
 }
 
 void set_integer_position(sf::Text &text,
                           const sf::Vector2f &position)
 {
-    text.setPosition(static_cast<int>(position.x), static_cast<int>(position.y));
+    // Cast to integer, then back to float
+    text.setPosition(static_cast<float>(static_cast<int>(position.x)),
+                     static_cast<float>(static_cast<int>(position.y)));
 }
 
 }  // namespace core::text
